Move parsing of commandline parameter out of CommandLineParameters into the CommandLineParser class
