+++ src/ChangeLog.Test/CommandLineParametersTest.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using CommandLine;
-using Grynwald.ChangeLog.Configuration;
 using Grynwald.Utilities.Configuration;
 using Xunit;
 
@@ -65,7 +64,7 @@ namespace Grynwald.ChangeLog.Test
         [Theory]
         [InlineData(null)]
         [InlineData("")]
-        public void CommandLineParameters_returns_null_is_output_path_is_null_or_empty(string outputPath)
+        public void CommandLineParameters_returns_null_if_output_path_is_null_or_empty(string outputPath)
         {
             // ARRANGE
             var sut = new CommandLineParameters()
@@ -77,13 +76,11 @@ namespace Grynwald.ChangeLog.Test
             Assert.Null(sut.OutputPath);
         }
 
-
         [Theory]
         [MemberData(nameof(Properties))]
         public void Properties_have_a_Option_attribute(string propertyName)
         {
-            // all properties inCommandLineParameters should have a OptionAttribute
-            // so the value can be set by the commandline parser
+            // All properties in CommandLineParameters should have a OptionAttribute so the value can be set by the commandline parser
 
             var property = typeof(CommandLineParameters).GetProperty(propertyName)!;
 
@@ -103,55 +100,6 @@ namespace Grynwald.ChangeLog.Test
                 Assert.NotNull(property.GetGetMethod());
                 Assert.NotNull(property.GetSetMethod());
             }
-        }
-
-        [Fact]
-        public void Template_parameter_is_optional()
-        {
-            // ARRANGE
-            var args = new[] { "--repository", "some-path" };
-
-            // ACT
-            var result = CommandLineParameters.Parse(args);
-
-            // ASSERT
-            Assert.Equal(ParserResultType.Parsed, result.Tag);
-            Assert.Equal(typeof(CommandLineParameters), result.TypeInfo.Current);
-            result.WithParsed(parsed =>
-            {
-                Assert.Null(parsed.Template);
-            });
-        }
-
-        public static IEnumerable<object[]> TemplateNames()
-        {
-            foreach (var value in Enum.GetValues(typeof(ChangeLogConfiguration.TemplateName)).Cast<ChangeLogConfiguration.TemplateName>())
-            {
-                yield return new object[] { value.ToString(), value };
-                yield return new object[] { value.ToString().ToLower(), value };
-                yield return new object[] { value.ToString().ToUpper(), value };
-            }
-        }
-
-
-        [Theory]
-        [MemberData(nameof(TemplateNames))]
-        public void Template_parameter_is_parsed_correctly(string template, ChangeLogConfiguration.TemplateName expected)
-        {
-            // ARRANGE
-            var args = new[] { "--repository", "some-path", "--template", template };
-
-            // ACT
-            var result = CommandLineParameters.Parse(args);
-
-            // ASSERT
-            Assert.Equal(ParserResultType.Parsed, result.Tag);
-            Assert.Equal(typeof(CommandLineParameters), result.TypeInfo.Current);
-            result.WithParsed(parsed =>
-            {
-                Assert.NotNull(parsed.Template);
-                Assert.Equal(expected, parsed.Template);
-            });
         }
     }
 }
